fix: extract dowload qr code
