+++ apps/journeys-admin/src/components/Editor/Toolbar/Items/ShareItem/QrCodeDialog/DownloadQrCode/index.ts
@@ -1,0 +1,1 @@
+export { DownloadQrCode } from './DownloadQrCode'
