+++ apps/journeys-admin/src/components/Editor/Toolbar/Items/ShareItem/QrCodeDialog/QrCodeDialog.tsx
@@ -1,23 +1,14 @@
-import Button from '@mui/material/Button'
-import ButtonGroup from '@mui/material/ButtonGroup'
-import ClickAwayListener from '@mui/material/ClickAwayListener'
 import Divider from '@mui/material/Divider'
-import Grow from '@mui/material/Grow'
-import MenuItem from '@mui/material/MenuItem'
-import MenuList from '@mui/material/MenuList'
-import Paper from '@mui/material/Paper'
-import Popper from '@mui/material/Popper'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
 import { useTranslation } from 'next-i18next'
-import { enqueueSnackbar } from 'notistack'
 import { QRCodeCanvas } from 'qrcode.react'
-import { MouseEvent, ReactElement, useEffect, useState } from 'react'
+import { ReactElement, useEffect, useState } from 'react'
 
 import { Dialog } from '@core/shared/ui/Dialog'
-import ChevronDownIcon from '@core/shared/ui/icons/ChevronDown'
 
 import { CodeDestination } from './CodeDestination'
+import { DownloadQrCode } from './DownloadQrCode'
 import { ScanCount } from './ScanCount'
 
 interface QrCodeDialogProps {
@@ -33,59 +24,10 @@ export function QrCodeDialog({
 }: QrCodeDialogProps): ReactElement {
   const { t } = useTranslation('apps-journeys-admin')
   const [to, setTo] = useState(initialJourneyUrl ?? '')
-  const [showDownloadMenu, setShowDownloadMenu] = useState(false)
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null)
 
   useEffect(() => {
     setTo(initialJourneyUrl ?? '')
   }, [initialJourneyUrl])
-
-  function handleMenuClick(event: MouseEvent<HTMLElement>): void {
-    setAnchorEl(event.currentTarget)
-    setShowDownloadMenu(!showDownloadMenu)
-  }
-
-  function handleDownloadQrCode(type: 'png' | 'svg'): void {
-    const canvas = document.getElementById(
-      'qr-code-download'
-    ) as HTMLCanvasElement | null
-    if (canvas != null) {
-      try {
-        const pngUrl = canvas
-          .toDataURL('image/png')
-          .replace('image/png', 'image/octet-stream')
-        const downloadLink = document.createElement('a')
-        downloadLink.href = pngUrl
-        downloadLink.download = `qr-code.${type}`
-        document.body.appendChild(downloadLink)
-        downloadLink.click()
-        document.body.removeChild(downloadLink)
-      } catch (error) {
-        if (
-          error.message ===
-          "Failed to execute 'toDataURL' on 'HTMLCanvasElement': Tainted canvases may not be exported."
-        ) {
-          enqueueSnackbar('Error downloading, check CORS setting', {
-            variant: 'error',
-            preventDuplicate: true
-          })
-        } else {
-          enqueueSnackbar('Error downloading', {
-            variant: 'error',
-            preventDuplicate: true
-          })
-        }
-      }
-    }
-  }
-
-  async function handleCopyClick(): Promise<void> {
-    await navigator.clipboard.writeText(to)
-    enqueueSnackbar('Link copied', {
-      variant: 'success',
-      preventDuplicate: true
-    })
-  }
 
   function handleChangeTo(url: string): void {
     setTo(url)
@@ -139,97 +81,16 @@ export function QrCodeDialog({
             }}
           >
             <ScanCount />
-            <Stack
-              spacing={3}
-              sx={{
-                alignItems: { xs: 'center', sm: 'start' }
-              }}
+            <DownloadQrCode to={to} />
+            <Typography
+              variant="body2"
+              color="secondary.main"
+              sx={{ display: { xs: 'none', sm: 'flex' } }}
             >
-              <ButtonGroup
-                variant="contained"
-                sx={{
-                  borderRadius: 2,
-                  width: 200,
-                  height: 42,
-                  boxShadow: 'none',
-                  '.MuiButtonGroup-grouped': {
-                    borderColor: 'background.paper'
-                  }
-                }}
-              >
-                <Button
-                  fullWidth
-                  onClick={() => {
-                    handleDownloadQrCode('png')
-                    setShowDownloadMenu(false)
-                  }}
-                  sx={{
-                    backgroundColor: 'secondary.main',
-                    borderTopLeftRadius: 8,
-                    borderBottomLeftRadius: 8
-                  }}
-                >
-                  {t('Download PNG')}
-                </Button>
-                <Button
-                  onClick={(e) => {
-                    handleMenuClick(e)
-                  }}
-                  sx={{
-                    backgroundColor: 'secondary.main',
-                    borderTopRightRadius: 8,
-                    borderBottomRightRadius: 8
-                  }}
-                >
-                  <ChevronDownIcon />
-                </Button>
-              </ButtonGroup>
-              <Popper
-                sx={{ zIndex: 1 }}
-                open={showDownloadMenu}
-                anchorEl={anchorEl}
-                transition
-                disablePortal
-              >
-                {({ TransitionProps }) => (
-                  <Grow {...TransitionProps}>
-                    <Paper>
-                      <ClickAwayListener
-                        onClickAway={() => setShowDownloadMenu(false)}
-                      >
-                        <MenuList id="split-button-menu" autoFocusItem>
-                          <MenuItem
-                            onClick={() => {
-                              handleDownloadQrCode('svg')
-                              setShowDownloadMenu(false)
-                            }}
-                          >
-                            {t('Download SVG')}
-                          </MenuItem>
-                          <MenuItem
-                            onClick={() => {
-                              void handleCopyClick()
-                              setShowDownloadMenu(false)
-                            }}
-                          >
-                            {t('Copy Short Link')}
-                          </MenuItem>
-                        </MenuList>
-                      </ClickAwayListener>
-                    </Paper>
-                  </Grow>
-                )}
-              </Popper>
-              <Typography
-                variant="body2"
-                color="secondary.main"
-                sx={{ display: { xs: 'none', sm: 'flex' } }}
-              >
-                {t(
-                  'Here is your unique QR code that will direct people to your journey when scanned.'
-                )}
-              </Typography>
-            </Stack>
+              {t(
+                'Here is your unique QR code that will direct people to your journey when scanned.'
+              )}
+            </Typography>
           </Stack>
         </Stack>
         <Divider />
